+++ packages/bundle/.babelrc
@@ -7,6 +7,7 @@
     }
   },
   "plugins": [
+    "@babel/plugin-proposal-class-properties",
     "@babel/plugin-proposal-object-rest-spread",
     "@babel/plugin-transform-runtime",
     "babel-plugin-version-transform"
