+++ packages/bundle/src/adaptiveCards/Attachment/AdaptiveCardAttachment.js
