+++ packages/component/.babelrc
@@ -7,7 +7,6 @@
     }
   },
   "plugins": [
-    "@babel/proposal-class-properties",
     "@babel/proposal-object-rest-spread",
     "babel-plugin-version-transform"
   ],
