+++ packages/component/src/Styles/StyleSet/AudioCardAttachment.js
@@ -1,3 +1,7 @@
+// TODO: [P4] We are moving attachments related to Adaptive Cards out of "component"
+//       Later, we will rewrite these attachments without Adaptive Cards
+//       We are leaving the CSS here as-is for now
+
 export default function ({
   paddingRegular
 }) {
