+++ packages/component/src/Styles/adaptiveCardHostConfig.js
@@ -1,4 +1,3 @@
-import { HostConfig } from 'adaptivecards';
 import defaultStyleSetOptions from './defaultStyleSetOptions';
 // https://docs.microsoft.com/en-us/adaptive-cards/rendering-cards/host-config
 
@@ -8,7 +7,7 @@ export default ({
   cardEmphasisBackgroundColor,
   primaryFont,
   subtle
-} = defaultStyleSetOptions) => new HostConfig({
+} = defaultStyleSetOptions) => ({
   containerStyles: {
     default: {
       foregroundColors: {
