+++ packages/component/src/index.tsx
@@ -6,10 +6,19 @@ import Avatar from './Activity/Avatar';
 import Bubble from './Activity/Bubble';
 import CarouselLayout from './Activity/CarouselLayout';
 import ErrorBox from './ErrorBox';
+import Localize, { localize } from './Localization/Localize';
 import SendStatus, { connectSendStatus } from './Activity/SendStatus';
 import SpeakActivity, { connectSpeakActivity } from './Activity/Speak';
 import StackedLayout, { connectStackedLayout } from './Activity/StackedLayout';
 import Timestamp from './Activity/Timestamp';
+
+import AudioContent from './Attachment/AudioContent';
+import HTMLVideoContent from './Attachment/HTMLVideoContent';
+import ImageContent from './Attachment/ImageContent';
+import TextContent from './Attachment/TextContent';
+import VideoContent from './Attachment/VideoContent';
+import VimeoContent from './Attachment/VimeoContent';
+import YouTubeContent from './Attachment/YouTubeContent';
 
 import DictationInterims, { connectDictationInterims } from './SendBox/DictationInterims';
 import MicrophoneButton, { connectMicrophoneButton } from './SendBox/MicrophoneButton';
@@ -21,10 +30,10 @@ import UploadButton, { connectUploadButton } from './SendBox/UploadButton';
 import concatMiddleware from './Middleware/concatMiddleware';
 import connectToWebChat from './connectToWebChat';
 import Context from './Context';
-import createAdaptiveCardsAttachmentMiddleware from './Middleware/Attachment/createAdaptiveCardMiddleware';
 import createCoreActivityMiddleware from './Middleware/Activity/createCoreMiddleware';
 import createCoreAttachmentMiddleware from './Middleware/Attachment/createCoreMiddleware';
 import createStyleSet from './Styles/createStyleSet';
+import getTabIndex from './Utils/TypeFocusSink/getTabIndex';
 
 declare var VERSION
 
@@ -32,6 +41,16 @@ const version = VERSION
 
 const Components = {
   Composer,
+  Localize,
+
+  // Components for recomposing activities and attachments
+  AudioContent,
+  HTMLVideoContent,
+  ImageContent,
+  TextContent,
+  VideoContent,
+  VimeoContent,
+  YouTubeContent,
 
   // Components for recomposing transcript
   Avatar,
@@ -70,9 +89,10 @@ export {
   concatMiddleware,
   connectToWebChat,
   Context,
-  createAdaptiveCardsAttachmentMiddleware,
   createCoreActivityMiddleware,
   createCoreAttachmentMiddleware,
   createStyleSet,
+  getTabIndex,
+  localize,
   version
 }
